dampened the controller again
