+++ pi/src/controller/mod.rs
@@ -1,15 +1,22 @@
+mod switch;
+
 #[derive(Debug)]
-pub struct Controller {}
+pub struct Controller {
+    switch: DampenedSwitch,
+}
 
 impl Controller {
     pub fn new() -> Self {
-        Controller {}
+        Controller {
+            switch: DampenedSwitch::new(CONFIG.controller.switch_debounce_duration),
+        }
     }
 }
 
 use anyhow::Error;
+use switch::DampenedSwitch;
 
-use crate::{communication::Communication, state::State};
+use crate::{CONFIG, communication::Communication, state::State};
 
 impl Controller {
     pub async fn perform_action(
@@ -18,37 +25,11 @@ impl Controller {
         comm: &Communication,
     ) -> Result<(), Error> {
         let on = state.mining_condition();
-        state.flip_plug_switch(on, &comm).await
-    }
-}
 
-mod switch {
-    use std::time::{Duration, Instant};
-
-    #[derive(Debug)]
-    pub struct DampenedSwitch {
-        time_to_switch: Duration,
-        command: bool,
-        received_since: Instant,
-    }
-
-    impl DampenedSwitch {
-        pub fn new(time_to_switch: Duration) -> Self {
-            Self {
-                time_to_switch,
-                command: false,
-                received_since: Instant::now(),
-            }
+        if self.switch.perform(on) {
+            state.flip_plug_switch(on, &comm).await?;
         }
 
-        fn perform(&mut self, command: bool) -> bool {
-            if command != self.command {
-                self.command = command;
-                self.received_since = Instant::now();
-                false
-            } else {
-                Instant::now().duration_since(self.received_since) > self.time_to_switch
-            }
-        }
+        Ok(())
     }
 }
